BOT-2 cleanup, added utilities module to further organize code
